Reposition the top and information elements
The logo has a link to the main page, and the header elements were set
in the correct positions with flexbox properties.
The hero elements will wrap and they are centered.
The images in the information are centered and there is a gap between
them. Also their subtext is centered.